+++ tests/test_dictionary.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import pytest
 
 from grascii.dictionary import Dictionary
-from grascii.dictionary.build import DictionaryBuilder
+from grascii.dictionary.build import DictionaryBuilder, DictionaryOutputOptions
 from grascii.dictionary.common import (
     DictionaryAlreadyExists,
     DictionaryNotFound,
@@ -20,10 +20,6 @@ from grascii.dictionary.uninstall import uninstall_dictionary
 
 
 class TestDictionaryBuildWarnings(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        logging.getLogger("grascii.dictionary.build").setLevel(logging.CRITICAL)
-
     output_dir = Path("tests/dictionaries/outputs")
 
     def setUp(self):
@@ -38,91 +34,97 @@ class TestDictionaryBuildWarnings(unittest.TestCase):
 
     def test_uncertainty(self):
         builder = DictionaryBuilder(
-            infiles=[Path("tests/dictionaries/uncertainty.txt")],
-            output=self.output_dir,
             count_words=True,
         )
-        builder.build()
-        self.assertEqual(len(builder.warnings), 9)
-        self.assertEqual(len(builder.errors), 0)
-        for warning in builder.warnings:
+        summary = builder.build(
+            infiles=[Path("tests/dictionaries/uncertainty.txt")],
+            output=DictionaryOutputOptions(self.output_dir),
+        )
+        self.assertEqual(len(summary.warnings), 9)
+        self.assertEqual(len(summary.errors), 0)
+        for warning in summary.warnings:
             self.assertEqual(warning.level, logging.WARNING)
             self.assertRegex(warning.message, "Uncertainty")
-        entry_count = sum(val for val in builder.entry_counts.values())
+        entry_count = sum(val for val in summary.entry_counts.values())
         self.assertEqual(entry_count, 9)
 
     def test_count(self):
         builder = DictionaryBuilder(
-            infiles=[Path("tests/dictionaries/count.txt")],
-            output=self.output_dir,
             count_words=True,
         )
-        builder.build()
-        self.assertEqual(len(builder.warnings), 10)
-        self.assertEqual(len(builder.errors), 0)
-        for warning in builder.warnings:
+        summary = builder.build(
+            infiles=[Path("tests/dictionaries/count.txt")],
+            output=DictionaryOutputOptions(self.output_dir),
+        )
+        self.assertEqual(len(summary.warnings), 10)
+        self.assertEqual(len(summary.errors), 0)
+        for warning in summary.warnings:
             self.assertEqual(warning.level, logging.WARNING)
             self.assertRegex(warning.message, "Too many")
-        entry_count = sum(val for val in builder.entry_counts.values())
+        entry_count = sum(val for val in summary.entry_counts.values())
         self.assertEqual(entry_count, 12)
 
     def test_spelling(self):
         builder = DictionaryBuilder(
-            infiles=[Path("tests/dictionaries/spell.txt")],
-            output=self.output_dir,
             count_words=True,
             words_file=Path("tests/dictionaries/words.txt"),
         )
-        builder.build()
-        self.assertEqual(len(builder.warnings), 5)
-        self.assertEqual(len(builder.errors), 0)
-        for warning in builder.warnings:
+        summary = builder.build(
+            infiles=[Path("tests/dictionaries/spell.txt")],
+            output=DictionaryOutputOptions(self.output_dir),
+        )
+        self.assertEqual(len(summary.warnings), 5)
+        self.assertEqual(len(summary.errors), 0)
+        for warning in summary.warnings:
             self.assertEqual(warning.level, logging.WARNING)
             self.assertRegex(warning.message, "not in words")
-        entry_count = sum(val for val in builder.entry_counts.values())
+        entry_count = sum(val for val in summary.entry_counts.values())
         self.assertEqual(entry_count, 9)
 
     def test_too_few_words(self):
         builder = DictionaryBuilder(
-            infiles=[Path("tests/dictionaries/too_few_words.txt")],
-            output=self.output_dir,
             count_words=True,
         )
-        builder.build()
-        self.assertEqual(len(builder.warnings), 0)
-        self.assertEqual(len(builder.errors), 9)
-        for error in builder.errors:
+        summary = builder.build(
+            infiles=[Path("tests/dictionaries/too_few_words.txt")],
+            output=DictionaryOutputOptions(self.output_dir),
+        )
+        self.assertEqual(len(summary.warnings), 0)
+        self.assertEqual(len(summary.errors), 9)
+        for error in summary.errors:
             self.assertEqual(error.level, logging.ERROR)
             self.assertRegex(error.message, "Too few")
-        entry_count = sum(val for val in builder.entry_counts.values())
+        entry_count = sum(val for val in summary.entry_counts.values())
         self.assertEqual(entry_count, 0)
 
     def test_not_grascii(self):
         builder = DictionaryBuilder(
-            infiles=[Path("tests/dictionaries/not_grascii.txt")],
-            output=self.output_dir,
             parse=True,
         )
-        builder.build()
-        self.assertEqual(len(builder.warnings), 0)
-        self.assertEqual(len(builder.errors), 10)
-        for error in builder.errors:
+        summary = builder.build(
+            infiles=[Path("tests/dictionaries/not_grascii.txt")],
+            output=DictionaryOutputOptions(self.output_dir),
+        )
+        self.assertEqual(len(summary.warnings), 0)
+        self.assertEqual(len(summary.errors), 10)
+        for error in summary.errors:
             self.assertEqual(error.level, logging.ERROR)
             self.assertRegex(error.message, "Failed to parse")
-        entry_count = sum(val for val in builder.entry_counts.values())
+        entry_count = sum(val for val in summary.entry_counts.values())
         self.assertEqual(entry_count, 1)
 
     def test_comments_and_whitespace(self):
-        builder = DictionaryBuilder(
-            infiles=[Path("tests/dictionaries/comments.txt")], output=self.output_dir
+        builder = DictionaryBuilder()
+        summary = builder.build(
+            infiles=[Path("tests/dictionaries/comments.txt")],
+            output=DictionaryOutputOptions(self.output_dir),
         )
-        builder.build()
-        self.assertEqual(len(builder.warnings), 6)
-        self.assertEqual(len(builder.errors), 0)
-        for warning in builder.warnings:
+        self.assertEqual(len(summary.warnings), 6)
+        self.assertEqual(len(summary.errors), 0)
+        for warning in summary.warnings:
             self.assertEqual(warning.level, logging.WARNING)
             self.assertRegex(warning.message, "Uncertainty")
-        entry_count = sum(val for val in builder.entry_counts.values())
+        entry_count = sum(val for val in summary.entry_counts.values())
         self.assertEqual(entry_count, 7)
 
 
@@ -134,11 +136,9 @@ class TestBuiltins(unittest.TestCase):
 
     def test_preanniversary(self):
         inputs = Path("dictionaries/builtins/preanniversary/").glob("*.txt")
-        builder = DictionaryBuilder(
-            infiles=inputs, check_only=True, count_words=True, parse=True
-        )
-        builder.build()
-        self.assertEqual(len(builder.errors), 0)
+        builder = DictionaryBuilder(count_words=True, parse=True)
+        summary = builder.build(inputs, None)
+        self.assertEqual(len(summary.errors), 0)
 
 
 @pytest.fixture
@@ -155,10 +155,11 @@ def tmp_dict_path(tmp_path, monkeypatch):
 @pytest.fixture(scope="module")
 def tmp_build_path(tmp_path_factory):
     build_path = tmp_path_factory.mktemp("search", numbered=False)
-    builder = DictionaryBuilder(
-        infiles=Path("tests/dictionaries/search.txt"), output=build_path
+    builder = DictionaryBuilder()
+    builder.build(
+        infiles=Path("tests/dictionaries/search.txt"),
+        output=DictionaryOutputOptions(build_path),
     )
-    builder.build()
     return build_path